ID-B-21

+++ PetFamily.Backend/src/PetFamily.Application/Volunteers/Pets/Commands/RemoveFilePet/RemoveFilePetHandler.cs
@@ -1,29 +1,36 @@
-using CSharpFunctionalExtensions;
-using Microsoft.AspNetCore.Server.HttpSys;
+using FluentValidation;
 using Microsoft.Extensions.Logging;
 using PetFamily.Application.Abstractions;
 using PetFamily.Application.FileProvider;
 using PetFamily.Application.Volunteers.Pets.Commands.AddPet;
 using PetFamily.Domain.Models.VO;
-using PetFamily.Domain.Shared;
 using FileInfo = PetFamily.Application.FileProvider.FileInfo;
 
-namespace PetFamily.Application.Volunteers.Pets.Commands.RemovePet;
+namespace PetFamily.Application.Volunteers.Pets.Commands.RemoveFilePet;
 
-public record RemovePetQuery(string FilePath, string BucketName) : IQuery;
-public class RemovePetHandler : IQueryHandler<string, RemovePetQuery>
+public record RemoveFilePetQuery(string FilePath, string BucketName) : IQuery;
+public class RemoveFilePetHandler : IQueryHandler<string, RemoveFilePetQuery>
 {
     private readonly IFileProvider _fileProvider;
     private readonly ILogger<AddPetHandler> _logger;
+    private readonly IValidator<RemoveFilePetQuery> _validator;
 
-    public RemovePetHandler(IFileProvider fileProvider, ILogger<AddPetHandler> logger)
+    public RemoveFilePetHandler(
+        IFileProvider fileProvider, 
+        ILogger<AddPetHandler> logger, 
+        IValidator<RemoveFilePetQuery> validator)
     {
         _fileProvider = fileProvider;
         _logger = logger;
+        _validator = validator;
     }
 
-    public async Task<string> Handle(RemovePetQuery query, CancellationToken cancellationToken)
+    public async Task<string> Handle(RemoveFilePetQuery query, CancellationToken cancellationToken)
     {
+        var validationResult = await _validator.ValidateAsync(query, cancellationToken);
+        if (!validationResult.IsValid)
+            return "";
+        
         var filePath = new FilePath(query.FilePath);
         var fileData = new FileInfo(filePath, query.BucketName);
         var deleteFile = await _fileProvider.DeleteFile(fileData, cancellationToken);
